Some tidy up tweaks
