+++ ompi/mca/coll/libpnbc_osc/Makefile.am
@@ -25,6 +25,7 @@
 sources = \
 	coll_libpnbc_osc.h \
 	coll_libpnbc_osc_component.c \
+	pnbc_osc_schedule.c \
 	pnbc_osc.c \
 	pnbc_osc_action_put.c \
 	pnbc_osc_trigger.c \
